Move oxygen factory pressure controls to subclass
