+++ src/js/buildings/OxygenFactory.js
@@ -66,6 +66,78 @@ class OxygenFactory extends Building {
         dampingFactor * (targetProductivity - this.productivity);
     }
   }
+
+  initUI(autoBuildContainer, cache) {
+    if (!autoBuildContainer || !cache) return;
+
+    const pressureControl = document.createElement('div');
+    pressureControl.classList.add('o2-pressure-control');
+    pressureControl.style.display = this.isBooleanFlagSet('terraformingBureauFeature')
+      ? 'flex'
+      : 'none';
+
+    const pressureCheckbox = document.createElement('input');
+    pressureCheckbox.type = 'checkbox';
+    pressureCheckbox.classList.add('o2-pressure-checkbox');
+    pressureCheckbox.checked = oxygenFactorySettingsRef.autoDisableAbovePressure;
+    pressureCheckbox.addEventListener('change', () => {
+      oxygenFactorySettingsRef.autoDisableAbovePressure = pressureCheckbox.checked;
+    });
+    pressureControl.appendChild(pressureCheckbox);
+
+    const pressureLabel = document.createElement('span');
+    pressureLabel.textContent = 'Disable if O2 P > ';
+    pressureControl.appendChild(pressureLabel);
+
+    const pressureInput = document.createElement('input');
+    pressureInput.type = 'number';
+    pressureInput.step = 1;
+    pressureInput.classList.add('o2-pressure-input');
+    pressureControl.appendChild(pressureInput);
+
+    const unitSpan = document.createElement('span');
+    unitSpan.classList.add('o2-pressure-unit');
+    unitSpan.textContent = 'kPa';
+    pressureControl.appendChild(unitSpan);
+
+    const update = () => {
+      if (document.activeElement !== pressureInput) {
+        pressureInput.value = oxygenFactorySettingsRef.disablePressureThreshold;
+      }
+    };
+    update();
+
+    pressureInput.addEventListener('input', () => {
+      const val = parseFloat(pressureInput.value);
+      oxygenFactorySettingsRef.disablePressureThreshold = val;
+    });
+
+    autoBuildContainer.appendChild(pressureControl);
+
+    cache.o2 = {
+      container: pressureControl,
+      checkbox: pressureCheckbox,
+      input: pressureInput,
+      unitSpan: unitSpan
+    };
+  }
+
+  updateUI(elements) {
+    const o2Els = elements?.o2;
+    if (!o2Els || !o2Els.container) return;
+
+    const enabled = this.isBooleanFlagSet('terraformingBureauFeature');
+    o2Els.container.style.display = enabled ? 'flex' : 'none';
+    if (o2Els.checkbox) {
+      o2Els.checkbox.checked = oxygenFactorySettingsRef.autoDisableAbovePressure;
+    }
+    if (o2Els.input && document.activeElement !== o2Els.input) {
+      o2Els.input.value = oxygenFactorySettingsRef.disablePressureThreshold;
+    }
+    if (o2Els.unitSpan) {
+      o2Els.unitSpan.textContent = 'kPa';
+    }
+  }
 }
 
 if (typeof module !== 'undefined' && module.exports) {
